refactor(filters): Rename substitutions mod to redactions
